+++ tests/testthat/setup.R
@@ -1,0 +1,2 @@
+if (!dir.exists("testdata"))
+  utils::unzip("testdata.zip", exdir = ".")